refactor out packetflag and recvpackets
